+++ docs/whats-new.rst
@@ -4,8 +4,8 @@ What's New
 v0.1.5 (???)
 ------------
 
+- Adds new logger that leverages on built-in python logging.
 - Changed license to MIT
-
 
 
 v0.1.4 (22 Oct 2015)